+++ CHANGELOG.md
@@ -12,7 +12,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - This change log.
 - Issue templates.
 - Dedicated installer.
-- Ability to query the file path of the global configuration file via the 'config-path' command.
+- Ability to query the file path of the global configuration file via the 'path config' command.
 - Feature flags to various crates.
 - Type aliases `Size`, `SignedSize`, `Offset`, and `UnsignedOffset`.
 
@@ -21,6 +21,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Switch to GNU GPLv3 license.
 - Reconfigure interpreter to accept a seaside executable file instead of a project directory.
 - Place unimplemented functionality behind a feature gate instead of commenting it out.
+- Replace 'exe-path' command with 'path binary' command.
 - Rename `runtime_data` table in `memory_map.segments` to `stack_and_heap`.
 - Rename `LexError` variant of `seaside_assembler::error::ParseError` to `Lex`.
 - Rename `number_fmt` module in `seaside_constants` crate to `number_format`.